feat(combat): A1b — runPlayerTurn reads effectiveDamageStatsOf (byte-identical); keep critBuff gate-partials
